Print cudnn version at the same time as cnmem usage
